Installing passenger gem on app.sh to install nginx

+++ scripts/app.sh
@@ -1,19 +1,18 @@
 # Ensure nginx is installed
-test -e /usr/local/nginx || /usr/local/bin/passenger-install-nginx-module --auto --auto-download \
-    --prefix /usr/local/nginx --extra-configure-flags='--with-http_gzip_static_module --with-http_ssl_module --with-http_stub_status_module'
+test -e /usr/local/nginx || {sudo gem install passenger; sudo /usr/local/bin/passenger-install-nginx-module --auto --auto-download \
+    --prefix /usr/local/nginx --extra-configure-flags='--with-http_gzip_static_module --with-http_ssl_module --with-http_stub_status_module'}
+
+test -e /usr/local/nginx || { echo "nginx not properly installed" >&2; exit 1; }
 
 LINE='export PATH=$PATH:/usr/local/nginx/sbin'
 grep -q "$LINE" .bashrc 2>/dev/null || echo "$LINE" >> .bashrc
 
-mkdir -p /usr/local/nginx/conf/sites-enabled /var/log/nginx/
-mkdir -p /var/log/nginx
+sudo mkdir -p /usr/local/nginx/conf/sites-enabled /var/log/nginx/
+sudo ruby /home/vagrant/scripts/configure_nginx.rb
 
-ruby /vagrant/scripts/configure_nginx.rb
-
-cd /home/ubuntu/documentcloud
-
+cd /home/vagrant/documentcloud
 test -e secrets/keys || { cp -r config/server/keys secrets/ ; }
 
-cp config/server/nginx/nginx.init /etc/init.d/nginx
-update-rc.d nginx defaults
-/etc/init.d/nginx start
+sudo cp config/server/nginx/nginx.init /etc/init.d/nginx
+sudo update-rc.d nginx defaults
+sudo /etc/init.d/nginx start
